# internal/app/api/core/assets/doc/v0_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WireGuard Portal API - UI Endpoints",
        "title": "WireGuard Portal SPA-UI API",
        "contact": {
            "name": "WireGuard Portal Developers",
            "url": "https://github.com/fedor-git/wg-portal-2"
        },
        "version": "0.0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/hostname": {
            "get": {
                "description": "Nothing more to describe...",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Testing"
                ],
                "summary": "Get the current host name.",
                "operationId": "test_handleHostnameGet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/now": {
            "get": {
                "description": "Nothing more to describe...",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Testing"
                ],
                "summary": "Get the current local time.",
                "operationId": "test_handleCurrentTimeGet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Message": {
                    "type": "string"
                }
            }
        },
...
    }
}